set up decision tree
Add interfaces and abstract classes to enable the creation of decision
trees. Set up the Agent class to initialize the A-Star tree and decision
tree correctly, and recursively walk down the decision tree to get an
action. Add one simple decision node and two simple action nodes to test
the tree.